fix(coding): add shared status color map
